+++ src/inspeqtor/experimental/optimize.py
@@ -8,11 +8,13 @@ import optax  # type: ignore
 from alive_progress import alive_it  # type: ignore
 
 import chex
+import jaxtyping
 
-from .utils import dataloader, create_step
+from .utils import dataloader
 from .model import LossMetric, load_model
 
 
+@jax.tree_util.register_dataclass
 @dataclass
 class DataBundled:
     control_params: jnp.ndarray
@@ -89,6 +91,60 @@ class HistoryEntryV3:
     aux: dict[str, jnp.ndarray]
 
 
+def create_step(
+    optimizer: optax.GradientTransformation,
+    loss_fn: (
+        typing.Callable[..., jnp.ndarray]
+        | typing.Callable[..., typing.Tuple[jnp.ndarray, typing.Any]]
+    ),
+    has_aux: bool = False,
+):
+    """The create_step function creates a training step function and a test step function.
+
+    loss_fn should have the following signature:
+    ```py
+    def loss_fn(params: jaxtyping.PyTree, *args) -> jnp.ndarray:
+        ...
+        return loss_value
+    ```
+    where `params` is the parameters to be optimized, and `args` are the inputs for the loss function.
+
+    Args:
+        optimizer (optax.GradientTransformation): `optax` optimizer.
+        loss_fn (typing.Callable[[jaxtyping.PyTree, ...], jnp.ndarray]): Loss function, which takes in the model parameters, inputs, and targets, and returns the loss value.
+        has_aux (bool, optional): Whether the loss function return aux data or not. Defaults to False.
+
+    Returns:
+        __type__: train_step, test_step
+    """
+
+    # * Generalized training step
+    @jax.jit
+    def train_step(
+        params: jaxtyping.PyTree,
+        optimizer_state: optax.OptState,
+        *args,
+        **kwargs,
+    ):
+        loss_value, grads = jax.value_and_grad(loss_fn, has_aux=has_aux)(
+            params, *args, **kwargs
+        )
+        updates, opt_state = optimizer.update(grads, optimizer_state, params)
+        params = optax.apply_updates(params, updates)
+
+        return params, opt_state, loss_value
+
+    @jax.jit
+    def test_step(
+        params: jaxtyping.PyTree,
+        *args,
+        **kwargs,
+    ):
+        return loss_fn(params, *args, **kwargs)
+
+    return train_step, test_step
+
+
 def train_model(
     # Random key
     key: jnp.ndarray,
@@ -105,6 +161,9 @@ def train_model(
     callbacks: list[typing.Callable] = [],
     # Number of epochs
     NUM_EPOCH: int = 1_000,
+    # Optional state
+    model_params: typing.Any = None,
+    opt_state: typing.Any = None,
 ):
     """Train the BlackBox model
 
@@ -154,9 +213,13 @@ def train_model(
 
     BATCH_SIZE = val_p.shape[0]
 
-    # Initialize the model parameters
-    model_params = model.init(init_key, train_p[0])
-    opt_state = optimizer.init(model_params)
+    if model_params is None:
+        # Initialize the model parameters if it is None
+        model_params = model.init(init_key, train_p[0])
+
+    if opt_state is None:
+        # Initalize the optimizer state if it is None
+        opt_state = optimizer.init(model_params)
 
     # histories: list[dict[str, typing.Any]] = []
     histories: list[HistoryEntryV3] = []
@@ -231,232 +294,3 @@ def clean_history_entries(
     # Transform the keys of the dictionary
     clean_histories = list(map(transform_key, clean_histories))
     return clean_histories
-
-
-# def default_trainable_v4(
-#     control_sequence: ControlSequence,
-#     metric: LossMetric,
-#     experiment_identifier: str,
-#     hamiltonian: typing.Callable | str,
-#     construct_model_fn: typing.Callable[
-#         [dict[str, int]], tuple[nn.Module, dict[str, typing.Any]]
-#     ],
-#     calculate_metrics_fn: typing.Callable,
-#     NUM_EPOCH: int = 1000,
-#     CHECKPOINT_EVERY: int = 100,
-# ):
-#     """Create trainable function for `ray.tune` for hyperparameter tuning
-
-#     Args:
-#         control_sequence (PulseSequence): Pulse sequence of dataset
-#         metric (LossMetric): Metric to be minimized for.
-#         experiment_identifier (str): The experiment identifier
-#         hamiltonian (typing.Callable | str): Ideal Hamiltonian function or name.
-#         model_choice (type[nn.Module], optional): Choice of the Blackbox model. Defaults to BasicBlackBoxV2.
-#         NUM_EPOCH (int, optional): Number of training epoch. Defaults to 1000.
-#         CHECKPOINT_EVERY (int, optional): Checkpointing every given number. Defaults to 100.
-
-#     Returns:
-#         typing.Callable: Trainable function that recieve hyperparameter configutation, dataset and random key.
-#     """
-#     from ray import train
-
-#     def trainable(
-#         config: dict[str, int],
-#         train_data: DataBundled,
-#         val_data: DataBundled,
-#         test_data: DataBundled,
-#         train_key: jnp.ndarray,
-#     ):
-#         optimizer = get_default_optimizer(8 * NUM_EPOCH)
-
-#         model, model_config = construct_model_fn(config)
-
-#         partial_loss_fn = partial(
-#             loss_fn,
-#             model=model,
-#             loss_metric=metric,
-#             calculate_metrics_fn=calculate_metrics_fn,
-#         )
-
-#         def prepare_report(history: list[HistoryEntryV3]):
-#             metric_types = [LossMetric.MSEE, LossMetric.AEF, LossMetric.WAEE]
-#             metrics = {}
-#             for entry in history:
-#                 for metric_type in metric_types:
-#                     metrics[f"{entry.loop}/{metric_type}"] = entry.aux[
-#                         metric_type
-#                     ].item()
-
-#             return metrics
-
-#         def callback(
-#             model_params: VariableDict,
-#             opt_state: optax.OptState,
-#             history: list[HistoryEntryV3],
-#         ) -> None:
-#             # Get the lasted 3 entries
-#             last_entries = history[-3:]
-
-#             loops = ["train", "val", "test"]
-#             # assert that the last 3 entries are from train, val, and test
-#             assert all(entry.loop in loops for entry in last_entries)
-
-#             # Prepare the report
-#             metrics = prepare_report(history)
-
-#             # Check if last_entry.step is divisible by 100
-#             if (last_entries[-1].step + 1) % CHECKPOINT_EVERY == 0:
-#                 # Checkpoint the model
-
-#                 # Clean the history entries
-#                 clean_histories = clean_history_entries(history)
-
-#                 with tempfile.TemporaryDirectory() as tmpdir:
-#                     _ = save_model(
-#                         path=tmpdir,
-#                         experiment_identifier=experiment_identifier,
-#                         control_sequence=control_sequence,
-#                         hamiltonian=hamiltonian,
-#                         model_config=model_config,
-#                         model_params=model_params,
-#                         history=clean_histories,
-#                         with_auto_datetime=False,
-#                     )
-
-#                     # Report the loss and val_loss to tune
-#                     train.report(
-#                         metrics=metrics,
-#                         checkpoint=train.Checkpoint.from_directory(tmpdir),
-#                     )
-#             else:
-#                 # Report the loss and val_loss to tune
-#                 train.report(
-#                     metrics=metrics,
-#                 )
-
-#             return None
-
-#         _, _, history = train_model(
-#             key=train_key,
-#             train_data=train_data,
-#             val_data=val_data,
-#             test_data=test_data,
-#             model=model,
-#             optimizer=optimizer,
-#             loss_fn=partial_loss_fn,
-#             NUM_EPOCH=NUM_EPOCH,
-#             callbacks=[callback],
-#         )
-
-#         # Prepare the report
-#         metrics = prepare_report(history[-3:])
-
-#         return metrics
-
-#     return trainable
-
-
-# def sample_from_search_space(search_space: typing.Mapping[str, Domain]):
-#     return {key: value.sample() for key, value in search_space.items()}
-
-
-# class SearchAlgo(StrEnum):
-#     HYPEROPT = "hyperopt"
-#     OPTUNA = "optuna"
-
-
-# def hypertuner(
-#     trainable: typing.Callable,
-#     train_data: DataBundled,
-#     test_data: DataBundled,
-#     val_data: DataBundled,
-#     train_key: jnp.ndarray,
-#     metric: LossMetric,
-#     search_space: typing.Mapping[str, Domain],
-#     num_samples: int = 100,
-#     search_algo: SearchAlgo = SearchAlgo.HYPEROPT,
-# ):
-#     """Perform hyperparameter tuning
-
-#     Args:
-#         trainable (typing.Callable): Trainable function
-#         train_pulse_parameters (jnp.ndarray): Training pulse parameters
-#         train_unitaries (jnp.ndarray): Training ideal unitary matrix
-#         train_expectation_values (jnp.ndarray): Training experiment expectation value
-#         test_pulse_parameters (jnp.ndarray): Testing pulse parameters
-#         test_unitaries (jnp.ndarray): Testing ideal unitary matrix
-#         test_expectation_values (jnp.ndarray): Testing experiment expectation value
-#         val_pulse_parameters (jnp.ndarray): Validating pulse parameters
-#         val_unitaries (jnp.ndarray): Validating ideal unitary matrix
-#         val_expectation_values (jnp.ndarray): Validating experiment expectation value
-#         train_key (jnp.ndarray): Random key
-#         metric (LossMetric): Metric to optimized for.
-#         num_samples (int, optional): The number of random configuration of hyperparameter. Defaults to 100.
-#         search_algo (SearchAlgo, optional): The search algorithm to be used for optimization. Defaults to SearchAlgo.HYPEROPT.
-#         search_spaces (_type_, optional): Search space of hyperparameters. Defaults to { "hidden_layer_1_1": (5, 50), "hidden_layer_1_2": (5, 50), "hidden_layer_2_1": (5, 50), "hidden_layer_2_2": (5, 50), }.
-#         initial_config (_type_, optional): Initial hyperparameters. Defaults to { "hidden_layer_1_1": 10, "hidden_layer_1_2": 20, "hidden_layer_2_1": 10, "hidden_layer_2_2": 20, }.
-
-#     Returns:
-#         _type_: Optimization result.
-#     """
-#     from ray import tune
-#     from ray.tune.search.hyperopt import HyperOptSearch
-#     from ray.tune.search.optuna import OptunaSearch
-#     from ray.tune.search import Searcher
-
-#     current_best_params = [{key: value.sample() for key, value in search_space.items()}]
-
-#     # Prepend 'val/' to the metric
-#     prepended_metric = f"val/{metric}"
-
-#     if search_algo == SearchAlgo.HYPEROPT:
-#         search_algo_instance: Searcher = HyperOptSearch(
-#             metric=prepended_metric,
-#             mode="min",
-#             points_to_evaluate=current_best_params,
-#         )
-#     elif search_algo == SearchAlgo.OPTUNA:
-#         search_algo_instance = OptunaSearch(
-#             metric=prepended_metric,
-#             mode="min",
-#         )
-
-#     run_config = tune.RunConfig(
-#         name="tune_experiment",
-#         checkpoint_config=tune.CheckpointConfig(
-#             num_to_keep=10,
-#         ),
-#     )
-
-#     tuner = tune.Tuner(
-#         tune.with_parameters(
-#             trainable,
-#             train_data=train_data,
-#             val_data=val_data,
-#             test_data=test_data,
-#             train_key=train_key,
-#         ),
-#         tune_config=tune.TuneConfig(
-#             search_alg=search_algo_instance,
-#             metric=prepended_metric,
-#             mode="min",
-#             num_samples=num_samples,
-#         ),
-#         param_space=search_space,  # type: ignore
-#         run_config=run_config,
-#     )
-
-#     results = tuner.fit()
-
-#     return results
-
-
-def get_best_hypertuner_results(results, metric: LossMetric, loop: str = "val"):
-    prepended_metric = f"{loop}/{metric}"
-
-    with results.get_best_result(
-        metric=prepended_metric, mode="min"
-    ).checkpoint.as_directory() as checkpoint_dir:
-        model_state, hist, data_config = load_model(checkpoint_dir, skip_history=False)
-    return model_state, hist, data_config
